Remove unnecessary thread wrapper class
